+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased][unreleased]
+
+- Fix Amazon Linux support
+
 ## [3.5.0] - 2015-12-21
 ### Added
 - Systemd support